Refactored most of the code from GeoAlignment into GeoAlignmentBase. Added a new HeadingAlignment that only changes the rotation to match the heading. Made GeoAlignment inherit from HeadingAlignment.
