+++ SpatialAlignment-Unity/Assets/SpatialAlignment-Examples/Geocentric/Scenes/SA-Geocentric.unity
@@ -411,7 +411,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 8
+  m_RootOrder: 9
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1468509572
 GameObject:
@@ -444,8 +444,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 43de0081f68fbe24c8c8f74ad956405c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  altitude: 0
   geoReference: {fileID: 231975971}
+  heading: 0
+  altitude: 0
   latitude: 30.04042
   longitude: -95.65007
   relativeAltitude: 1
@@ -564,8 +565,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 43de0081f68fbe24c8c8f74ad956405c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  altitude: 0
   geoReference: {fileID: 231975971}
+  heading: 0
+  altitude: 0
   latitude: 30.04093
   longitude: -95.65057
   relativeAltitude: 1
@@ -669,8 +671,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 43de0081f68fbe24c8c8f74ad956405c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  altitude: 0
   geoReference: {fileID: 231975971}
+  heading: 0
+  altitude: 0
   latitude: 30.04073
   longitude: -95.64962
   relativeAltitude: 1
@@ -807,6 +810,115 @@ MeshFilter:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2059821774}
   m_Mesh: {fileID: 10207, guid: 0000000000000000e000000000000000, type: 0}
+--- !u!1001 &2110799786
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 2126429526}
+    m_Modifications:
+    - target: {fileID: 100002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_Name
+      value: Direction
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -0.05
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0.4
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 90
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 05d0678d150ba6c409d42cde9979d052, type: 3}
+--- !u!4 &2110799787 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 400002, guid: 05d0678d150ba6c409d42cde9979d052,
+    type: 3}
+  m_PrefabInstance: {fileID: 2110799786}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1 &2126429524
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 2126429526}
+  - component: {fileID: 2126429525}
+  m_Layer: 0
+  m_Name: North
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &2126429525
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2126429524}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 25778b7d2a760fd409d3a9fb12d02f74, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  geoReference: {fileID: 231975971}
+  heading: 0
+--- !u!4 &2126429526
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2126429524}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children:
+  - {fileID: 2110799787}
+  m_Father: {fileID: 0}
+  m_RootOrder: 8
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &2130187286
 GameObject:
   m_ObjectHideFlags: 0
