+++ SpatialAlignment-Unity/Assets/SpatialAlignment/Strategies/Geocentric/GeoAlignmentBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aec9196469a941e4984d2d15033fb44b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
