+++ SpatialAlignment-Unity/Assets/SpatialAlignment/Strategies/Geocentric/HeadingAlignment.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25778b7d2a760fd409d3a9fb12d02f74
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
